Fixed null exception for icon and now creates .gitlab-ci-local folder automatically, if .gitlab-ci.yml exists
