+++ src/main/kotlin/dk/cego/gitlab_ci_local_plugin/GclToolWindow.form
@@ -20,8 +20,7 @@
               <grid row="0" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
             </constraints>
             <properties>
-              <icon value="actions/refresh.png"/>
-              <text value="Refresh"/>
+              <text value=""/>
             </properties>
           </component>
           <component id="93a33" class="javax.swing.JCheckBox" binding="checkBox">
